feat: new modular structure for components
