+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
-import { ChatFilledRefreshed, ChatRefreshed } from "./components/ui/icons";
+import { ChatFilledRefreshed, ChatRefreshed } from "./components/ui/Icons";
 
 function App() {
   const [count, setCount] = useState(0);
