+++ src/components/composites/sidebar/Sidebar.module.css
@@ -19,18 +19,6 @@
   flex-grow: 1;
 }
 
-.sidebarNav,
-.userOptions {
-  margin: 2px 0;
-  display: flex;
-  flex-wrap: nowrap;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  gap: 4px;
-  color: rgba(var(--nav-buttons-text-color), 0.6);
-}
-
 /* .sidebarNav > * {
   margin: 2px 0;
 } */
@@ -40,20 +28,4 @@
   width: 100%;
   height: auto;
   flex-shrink: 1;
-}
-
-.userOptions {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-.divider {
-  height: 1px;
-  width: 100%;
-  border: none;
-  border-top: solid 1px rgba(var(--nav-divider-icon-color), 0.1);
-  margin: 8px 0;
 }
