+++ src/components/composites/sidebar/components/NavButtons.module.css
@@ -1,0 +1,10 @@
+.navButtons {
+  margin: 2px 0;
+  display: flex;
+  flex-wrap: nowrap;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 4px;
+  color: rgba(var(--nav-buttons-text-color), 0.6);
+}
