+++ src/components/composites/sidebar/components/UserOptions.tsx
@@ -1,0 +1,21 @@
+import { RoundedButton } from "../../../ui/Buttons";
+import { SettingsRefreshed } from "../../../ui/Icons";
+import ProfileAvatar from "../../../ui/ProfileAvatar";
+
+import styles from "./UserOptions.module.css";
+
+export default function UserOptions() {
+  return (
+    <div>
+      <div className={styles.userOptions}>
+        <hr className={styles.divider} />
+        <RoundedButton>
+          <SettingsRefreshed pointerEvents={"none"} />
+        </RoundedButton>
+        <RoundedButton>
+          <ProfileAvatar size="small" url="/Avatars/image_0002.png" />
+        </RoundedButton>
+      </div>
+    </div>
+  );
+}
