+++ src/components/ui/buttons/index.tsx
@@ -1,1 +1,1 @@
-export { default as GenericButton } from "./RoundedButton";
+export { default as RoundedButton } from "./RoundedButton/RoundedButton";
